Corrections sur l'exercice 2, continuation de l'exercice 3

+++ Java_Atelier_1/Variables.java
@@ -2,11 +2,12 @@ package Java_Atelier_1;
 
 public class Variables {
 	public void Sqr_root() {
-		double a = 3.0;
+		// imprime la racine carrée de a * a + b * b
+		float a = (float) 3.0;
 		double b = 4;
-		double c;
+		float c;
 		
-		c = Math.sqrt((a * a) + (b * b));
+		c = (float) Math.sqrt((a * a) + (b * b));
 		System.out.println("c = " + c);
 	}
 }
